Deleted Test file i/o methods and Begin Adding file i/o methods to DataBaseAdapter.
Will handle current file(session aka csv) as a single ref within the DataBaseAdapter Class. The current design of the Db class has each method that deals with the csv file opening/closing the outputStream rather than ref a single ouputStream  i'm not sure if it matters but I can swap if it does.